# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open Social Care ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/password/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User reset password",
                "description": "User reset password",
                "operationId": "Password reset password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "description": "token",
                                        "type": "string",
                                        "default": "123456"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "password",
                                        "default": "123456"
                                    },
                                    "password_confirmation": {
                                        "description": "password_confirmation",
                                        "type": "password",
                                        "default": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset Successfully"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}